подготовка к введению управления частотой преобразователя подсветки
git-svn-id: http://dolphin/svn/svn2/trunk@468 953d4507-ba89-49e6-903b-66a71853bf49

+++ board.h
@@ -106,7 +106,7 @@ void board_set_dacscale(uint_fast8_t n);	/* Использование амплитуды сигнала с ЦА
 void board_set_dac1(uint_fast8_t n);	/* подстройка опорного генератора */
 void board_set_bglight(uint_fast8_t n);	/* Включение подсветки дисплея */
 void board_set_kblight(uint_fast8_t v);	/* Включение подсветки клавиатуры */
-void board_set_blfreq(uint_fast8_t n);	/* установка делителя для формирования рабочей частоты преобразователя подсветки */
+void board_set_blfreq(uint_fast32_t n);	/* установка делителя для формирования рабочей частоты преобразователя подсветки */
 void board_set_txcw(uint_fast8_t v);	/* Включение передачи в обход балансного модулятора */
 void board_set_txgate(uint_fast8_t v);	/* разрешение драйвера и оконечного усилителя */
 void board_set_scalelo1(uint_fast8_t n);	/* SW-2011-RDX */
